# resources/api-spec.yml
openapi: 3.0.3
info:
  title: Payment API
  version: 1.0.0
paths:
  /payment:
    post:
      summary: Оплата
      description: POST API запрос на совершение платежа
      operationId: makePayment
      tags:
        - Payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Платёж успешно прошёл
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Запрос не прошёл валидацию
        '500':
          description: Внутренняя ошибка сервера
  /balance:
    post:
      summary: Создание баланса
      description: POST API запрос на создание баланса
      operationId: createBalance
      tags:
        - Balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BalanceRequest'
      responses:
        '200':
          description: Баланс успешно создался
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '500':
          description: Внутренняя ошибка сервера
  /balance/{userId}:
    get:
      summary: Получение баланса
      description: GET API запрос на получение баланса
      operationId: getBalance
      tags:
        - Balance
      parameters:
        - name: userId
          in: path
          required: true
          description: Идентификатор баланса
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Баланс пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '500':
          description: Внутренняя ошибка сервера

security:
  - OAuth2: ["payment_read", "payment_write", "balance_read", "balance_write"]

components:
  securitySchemes:
    OAuthService:
      type: oauth2
      description: OAuth2 Client Credentials flow
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8100/realms/market-app/protocol/openid-connect/token
          refreshUrl: http://localhost:8100/realms/market-app/protocol/openid-connect/token
          scopes:
            payment_read: Чтение платежей
            payment_write: Создание платежей
            balance_read: Чтение баланса
            balance_write: Управление балансом
  schemas:
    BalanceRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          x-java-type: java.util.UUID
        balance:
          type: number
          format: decimal
          example: "100.50"
          x-java-type: java.math.BigDecimal
    PaymentRequest:
      type: object
      properties:
        balanceId:
          type: integer
          format: int64
        amount:
          type: number
          format: decimal
          example: "100.50"
          x-java-type: java.math.BigDecimal
      required:
        - amount
    PaymentResponse:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
          example: 1
        paymentDateTime:
          type: string
          format: date-time
          x-java-type: java.time.LocalDateTime
    BalanceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: string
          format: uuid
          x-java-type: java.util.UUID
        balance:
          type: number
          format: decimal
          example: "100.50"
          x-java-type: java.math.BigDecimal